refactor: rename module from container-data-collector to data-collector
