+++ metrics-backend/data-collector/Dockerfile
@@ -3,7 +3,7 @@ FROM gradle:8.5.0-jdk21 AS build
 WORKDIR /build
 
 # Gradle 캐시 최적화: 의존성 관련 파일만 먼저 복사
-COPY metrics-backend/container-data-collector/build.gradle metrics-backend/container-data-collector/
+COPY metrics-backend/data-collector/build.gradle metrics-backend/data-collector/
 COPY metrics-backend/build.gradle metrics-backend/
 COPY build.gradle settings.gradle ./
 COPY gradle gradle
@@ -15,7 +15,7 @@ COPY gradle/wrapper gradle/wrapper
 COPY . .
 
 # container-data-collector 모듈만 빌드
-RUN ./gradlew :metrics-backend:container-data-collector:clean :metrics-backend:container-data-collector:build -x test
+RUN ./gradlew :metrics-backend:data-collector:clean :metrics-backend:data-collector:build -x test
 
 RUN find /build -name "*.jar"
 
@@ -28,7 +28,7 @@ RUN adduser --disabled-password --gecos '' appuser \
   && chown -R appuser:appuser /app
 
 # 빌드된 JAR 복사
-COPY --from=build /build/metrics-backend/container-data-collector/build/libs/container-data-collector-0.0.1-SNAPSHOT.jar app.jar
+COPY --from=build /build/metrics-backend/data-collector/build/libs/data-collector-0.0.1-SNAPSHOT.jar app.jar
 
 USER appuser
