+++ metrics-backend/data-collector/src/main/java/kr/cs/interdata/datacollector/DockerStatsCollector.java
@@ -1,4 +1,4 @@
-package kr.cs.interdata.containerdatacollector;
+package kr.cs.interdata.datacollector;
 
 import com.github.dockerjava.api.DockerClient;
 import com.github.dockerjava.api.model.Container;
