+++ metrics-backend/data-collector/src/main/java/kr/cs/interdata/datacollector/MachineResourceMonitor.java
@@ -1,4 +1,4 @@
-package kr.cs.interdata.containerdatacollector;
+package kr.cs.interdata.datacollector;
 
 import com.google.gson.Gson;
 import oshi.SystemInfo;
